refactor:폴더 구조에 따른 파일 이동
