+++ src/components/auth/UserTypeSelector.tsx
@@ -1,4 +1,4 @@
-import type { UserType } from "../../../types/auth";
+import type { UserType } from "../../types/auth";
 
 interface UserTypeSelectorProps {
   value: UserType | '';
